+++ regression/reference/clibrary/pyClibrarymodule.c
@@ -245,6 +245,84 @@ PY_Function4a(
     return (PyObject *) SHTPy_rv;
 // splicer end function.function4a
 }
+
+static char PY_ImpliedLen__doc__[] =
+"documentation"
+;
+
+/**
+ * \brief Return the implied argument - text length
+ *
+ * Pass the Fortran length of the char argument directy to the C function.
+ * No need for the bufferify version which will needlessly copy the string.
+ */
+static PyObject *
+PY_ImpliedLen(
+  PyObject *SHROUD_UNUSED(self),
+  PyObject *args,
+  PyObject *kwds)
+{
+// int ImpliedLen(const char * text +intent(in), int ltext +implied(len(text))+intent(in)+value)
+// splicer begin function.implied_len
+    const char * text;
+    char *SHT_kwlist[] = {
+        "text",
+        NULL };
+
+    if (!PyArg_ParseTupleAndKeywords(args, kwds, "s:ImpliedLen",
+        SHT_kwlist, &text))
+        return NULL;
+
+    // pre_call
+    int ltext = strlen(text);
+
+    int SHC_rv = ImpliedLen(text, ltext);
+
+    // post_call
+    PyObject * SHTPy_rv = PyInt_FromLong(SHC_rv);
+
+    return (PyObject *) SHTPy_rv;
+// splicer end function.implied_len
+}
+
+static char PY_ImpliedLenTrim__doc__[] =
+"documentation"
+;
+
+/**
+ * \brief Return the implied argument - text length
+ *
+ * Pass the Fortran length of the char argument directy to the C function.
+ * No need for the bufferify version which will needlessly copy the string.
+ */
+static PyObject *
+PY_ImpliedLenTrim(
+  PyObject *SHROUD_UNUSED(self),
+  PyObject *args,
+  PyObject *kwds)
+{
+// int ImpliedLenTrim(const char * text +intent(in), int ltext +implied(len_trim(text))+intent(in)+value)
+// splicer begin function.implied_len_trim
+    const char * text;
+    char *SHT_kwlist[] = {
+        "text",
+        NULL };
+
+    if (!PyArg_ParseTupleAndKeywords(args, kwds, "s:ImpliedLenTrim",
+        SHT_kwlist, &text))
+        return NULL;
+
+    // pre_call
+    int ltext = strlen(text);
+
+    int SHC_rv = ImpliedLenTrim(text, ltext);
+
+    // post_call
+    PyObject * SHTPy_rv = PyInt_FromLong(SHC_rv);
+
+    return (PyObject *) SHTPy_rv;
+// splicer end function.implied_len_trim
+}
 static PyMethodDef PY_methods[] = {
 {"Function1", (PyCFunction)PY_Function1, METH_NOARGS,
     PY_Function1__doc__},
@@ -257,6 +335,10 @@ static PyMethodDef PY_methods[] = {
     PY_Function3b__doc__},
 {"Function4a", (PyCFunction)PY_Function4a, METH_VARARGS|METH_KEYWORDS,
     PY_Function4a__doc__},
+{"ImpliedLen", (PyCFunction)PY_ImpliedLen, METH_VARARGS|METH_KEYWORDS,
+    PY_ImpliedLen__doc__},
+{"ImpliedLenTrim", (PyCFunction)PY_ImpliedLenTrim,
+    METH_VARARGS|METH_KEYWORDS, PY_ImpliedLenTrim__doc__},
 {NULL,   (PyCFunction)NULL, 0, NULL}            /* sentinel */
 };