+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Added format dictionary field template_suffix.
+- Added the F_create_bufferify_function option.
 
 ### Fixed
 - C++ function arguments which pass a class by value.